requirements.txt
openai-whisper
googletrans==4.0.0-rc1
pydub
requests
sounddevice
numpy
scipy
python-dotenv